Incorporating some of Jordan's initial comments
